Role: refactor - unique form for role creation and edition (WIP)
