+++ lib/frontend/src/components/CreateRole.sass
@@ -8,43 +8,3 @@
   &__createRole
     margin-left: 1em
     margin-right: .5em
-
-  input, select
-    height: 2.5em
-
-  input, select, textarea
-    background: #e0e9f0
-    border: none
-    box-sizing: border-box
-    font-family: Lato, sans-serif
-    margin: 0
-    max-width: 100%
-    outline: none
-    width: 60em
-
-  label
-    display: block
-    margin: .7em 0
-    max-width: 100%
-    width: 50em
-
-  [type=submit], [type=button]
-    background: #bac9d5
-    border: none
-    font-family: $text-font
-    margin: .5em 0 .5em 1em
-    padding: 1em
-    transition: background 0.2s
-
-  &__duration
-    align-items: center
-    display: flex
-
-    input
-      width: 5em
-
-  &__endDate
-    color: #323232
-    font-size: .9em
-    font-style: italic
-    margin-left: 1em
