+++ lib/frontend/src/components/Role/Role.sass
@@ -54,23 +54,3 @@
 
   input, select
     height: 2.5em
-
-
-  &__duration
-    align-items: center
-    display: flex
-
-    input
-      width: 5em
-
-  &__endDate
-    color: #323232
-    font-size: .9em
-    font-style: italic
-    margin-left: 1em
-
-  &__long
-    width: 60em
-
-  &__short
-    width: 15em
