+++ lib/frontend/src/components/Role/index.jsx
@@ -1,6 +1,5 @@
 import './Role.sass'
 
-import { addDays, format } from 'date-fns'
 import React from 'react'
 import { Helmet } from 'react-helmet'
 import { withRouter } from 'react-router-dom'
@@ -12,13 +11,12 @@ import {
   fetchItems,
   fetchRole,
   fetchRoles,
-  updateRole,
 } from '../../actions/roles'
 import { block, connect } from '../../utils'
 import Items from './Items'
 import RoleFocus from './RoleFocus'
 import Loading from './../Loading'
-import MarkdownEditor from './../Utils/MarkdownEditor'
+import RoleForm from './../Utils/RoleForm'
 
 var ReactMarkdown = require('react-markdown')
 
@@ -52,15 +50,6 @@ class Role extends React.Component {
     }
   }
 
-  calculateEndDate(duration) {
-    this.setState({
-      endDate:
-        duration || duration !== 0
-          ? format(addDays(new Date(), duration), 'DD/MM/YYYY (dddd)')
-          : null,
-    })
-  }
-
   render() {
     const {
       btnClick,
@@ -70,12 +59,11 @@ class Role extends React.Component {
       history,
       items,
       loading,
-      onEditRole,
       onGoBack,
       role,
       roles,
+      users,
     } = this.props
-    const { endDate } = this.state
     return (
       <section className={b()}>
         <Helmet>
@@ -130,95 +118,13 @@ class Role extends React.Component {
           )}
           {loading && <Loading />}
           {role.is_in_edition ? (
-            <article>
-              <form
-                onSubmit={e => {
-                  e.preventDefault()
-                  onEditRole(role, e, history)
-                }}
-              >
-                <h1>Edit {role.role_name} role :</h1>
-                <label>
-                  Circle :
-                  <select
-                    name="circle_id"
-                    defaultValue={role.circle_id}
-                    className={b('long')}
-                  >
-                    {circles.map(circle => (
-                      <option key={circle.circle_id} value={circle.circle_id}>
-                        {circle.circle_name}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label>
-                  Name :
-                  <input
-                    name="role_name"
-                    className={b('long')}
-                    defaultValue={role.role_name}
-                    required
-                  />
-                </label>
-                <label>
-                  Type :
-                  <select
-                    name="role_type"
-                    defaultValue={role.role_type ? role.role_type : ''}
-                    required
-                    className={b('long')}
-                  >
-                    <option value="" />
-                    {roleTypes.map(roleType => (
-                      <option key={roleType.value} value={roleType.value}>
-                        {roleType.name}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label>
-                  Duration (in days) :
-                  <span className={b('duration')}>
-                    <input
-                      defaultValue={role.duration ? role.duration : ''}
-                      onChange={e => this.calculateEndDate(+e.target.value)}
-                      name="duration"
-                      type="number"
-                    />
-                    <span className={b('endDate')}>
-                      {endDate && `until ${endDate}`}
-                    </span>
-                  </span>
-                </label>
-                <label>
-                  Purpose :
-                  <input
-                    name="role_purpose"
-                    className={b('long')}
-                    defaultValue={role.role_purpose}
-                    required
-                  />
-                </label>
-                <label>
-                  Domain :
-                  <input
-                    name="role_domain"
-                    className={b('long')}
-                    defaultValue={role.role_domain}
-                    required
-                  />
-                </label>
-                <label>
-                  Accountabilities :
-                  <MarkdownEditor />
-                </label>
-                <button type="submit">Edit role</button>
-                <button type="submit" onClick={() => editClick()}>
-                  Cancel
-                </button>
-              </form>
-            </article>
+            <RoleForm
+              circle={{}}
+              circles={circles}
+              isCreation
+              role={role}
+              users={users}
+            />
           ) : (
             <article>
               <h3>Circle</h3>
@@ -289,6 +195,7 @@ export default withRouter(
       location: state.router.location,
       role: state.role.results,
       roles: state.roles.results,
+      users: state.users.results,
     }),
     dispatch => ({
       btnClick: (roleId, circleId, history) => {
@@ -297,22 +204,6 @@ export default withRouter(
       editClick: content => {
         dispatch(editRole())
         dispatch(updateMarkdown(content))
-      },
-      onEditRole: (role, e, history) => {
-        const formRole = [].slice.call(e.target.elements).reduce(
-          function(map, obj) {
-            if (obj.name === 'body') {
-              map.role_accountabilities = obj.value
-            } else if (obj.name) {
-              map[obj.name] = obj.value
-            }
-            return map
-          },
-          { circle_id: role.circle_id }
-        )
-        dispatch(updateRole(role.role_id, formRole, history))
-        dispatch(updateMarkdown(''))
-        dispatch(fetchRole())
       },
       onGoBack: (circleId, history) => {
         history.push(`/circle?circle_id=${circleId}`)
