+++ lib/frontend/src/components/Utils/RoleForm.sass
@@ -1,0 +1,52 @@
+@import ../vars
+
+.RoleForm
+  flex: 1
+  overflow: auto
+  position: relative
+
+  input, select
+    height: 2.5em
+
+  input, select, textarea
+    background: #e0e9f0
+    border: none
+    box-sizing: border-box
+    font-family: Lato, sans-serif
+    margin: 0
+    max-width: 100%
+    outline: none
+    width: 60em
+
+  label
+    display: block
+    margin: .7em 0
+    max-width: 100%
+    width: 50em
+
+  [type=submit], [type=button]
+    background: #bac9d5
+    border: none
+    font-family: $text-font
+    margin: .5em 0 .5em 1em
+    padding: 1em
+    transition: background 0.2s
+
+  &__duration
+    align-items: center
+    display: flex
+
+    input
+      width: 5em
+
+  &__endDate
+    color: #323232
+    font-size: .9em
+    font-style: italic
+    margin-left: 1em
+
+  &__long
+    width: 60em
+
+  &__short
+    width: 15em
